Fixing some amber warnings
Really small semantic stuff

+++ Endless RunnerTests/HighScoresViewControllerTest.m
@@ -34,7 +34,7 @@
     XCTAssertNotNil(self.app.highscores);
     bool highscoresaved;
     NSUInteger scoreIndex = [self.app.highscores indexOfObject: @"Jennifer"];
-    if (scoreIndex == nan) {
+    if (scoreIndex == (int)nan) {
         highscoresaved = false;
     } else {
         highscoresaved = true;
